requirements.txt
appdirs==1.4.4
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
docker-pycreds==0.4.0
et-xmlfile==1.1.0
filelock==3.13.1
fsspec==2023.12.2
gitdb==4.0.11
gitpython==3.1.41
idna==3.6
jinja2==3.1.3
joblib==1.3.2
markupsafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
openpyxl==3.1.2
pandas==2.0.3
pillow==10.2.0
pip==25.0.1
pkg-resources==0.0.0
protobuf==4.25.2
psutil==5.9.7
python-dateutil==2.8.2
pytz==2023.3.post1
pyyaml==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==1.39.2
setproctitle==1.3.3
setuptools==44.0.0
six==1.16.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.2
torchvision==0.16.2
triton==2.1.0
typing-extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wandb==0.16.2
wheel==0.45.1